Refactor simulation timeline series config
