+++ libs/@hashintel/petrinaut/src/ui/views/Editor/panels/BottomPanel/subviews/simulation-timeline/types.ts
@@ -1,8 +1,28 @@
+import type { SimulationFrameReader } from "../../../../../../../react/simulation/context";
+
+/** Frame reader consumed by timeline series extractors. */
+export type TimelineFrame = SimulationFrameReader;
+
 /** Metadata for each plotted series, stable across streaming updates. */
 export interface TimelineSeriesMeta {
   seriesId: string;
   seriesName: string;
   color: string;
+}
+
+/**
+ * Returns the value for series `seriesIdx` at the given frame.
+ * Returning NaN leaves a gap on the chart.
+ */
+export type TimelineSeriesExtractor = (
+  frame: TimelineFrame,
+  seriesIdx: number,
+) => number;
+
+/** Series metadata and extractor for the active timeline view. */
+export interface TimelineSeriesConfig {
+  series: TimelineSeriesMeta[];
+  extract: TimelineSeriesExtractor;
 }
 
 /**
